AxisV2_backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
gspread==6.0.0
google-auth==2.27.0
pydantic==2.5.3
openai==1.58.1
twilio==9.4.0
websockets==13.1
httpx==0.27.0
